Refactor table types under TableStyle in Objects
Move cell style, border, and related enums from Entities to TableStyle in Objects. Update all references and templates to use new structure for improved organization and clarity. No functional changes.
