--- src/ACadSharp/Entities/TableEntity.BorderType.cs
@@ -1,17 +1,0 @@
-﻿namespace ACadSharp.Entities
-{
-	public partial class TableEntity
-	{
-		public enum BorderType : short
-		{
-			/// <summary>
-			/// Single border line.
-			/// </summary>
-			Single = 1,
-			/// <summary>
-			/// Double border line.
-			/// </summary>
-			Double = 2
-		}
-	}
-}
