--- src/ACadSharp/Entities/TableEntity.CellAlignmentType.cs
@@ -1,19 +1,0 @@
-﻿namespace ACadSharp.Entities
-{
-	public partial class TableEntity
-	{
-		public enum CellAlignmentType
-		{
-			None = 0,
-			TopLeft = 1,
-			TopCenter = 2,
-			TopRight = 3,
-			MiddleLeft = 4,
-			MiddleCenter = 5,
-			MiddleRight = 6,
-			BottomLeft = 7,
-			BottomCenter = 8,
-			BottomRight = 9,
-		}
-	}
-}
