--- src/ACadSharp/Entities/TableEntity.CellStyleClass.cs
@@ -1,12 +1,0 @@
-﻿namespace ACadSharp.Entities
-{
-	public partial class TableEntity
-	{
-		public enum CellStyleClass
-		{
-			Data = 1,
-
-			Label = 2
-		}
-	}
-}
