+++ src/ACadSharp/Entities/TableEntity.ITableComponent.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using static ACadSharp.Objects.TableStyle;
 
 namespace ACadSharp.Entities;
