+++ src/ACadSharp/IO/DXF/DxfStreamReader/DxfObjectsSectionReader.cs
@@ -11,6 +11,7 @@ using System.IO;
 using System.Linq;
 using static ACadSharp.IO.Templates.CadEvaluationGraphTemplate;
 using static ACadSharp.IO.Templates.CadTableEntityTemplate;
+using static ACadSharp.IO.Templates.CadTableStyleTemplate;
 
 namespace ACadSharp.IO.DXF
 {
@@ -1389,7 +1390,7 @@ namespace ACadSharp.IO.DXF
 
 		private void readFormattedCellContent()
 		{
-			TableEntity.ContentFormat format = new();
+			TableStyle.ContentFormat format = new();
 			CellContentFormatTemplate template = new CellContentFormatTemplate(format);
 			var map = DxfClassMap.Create(format.GetType(), "FORMATTEDCELLCONTENT");
 
@@ -1493,31 +1494,31 @@ namespace ACadSharp.IO.DXF
 
 		private void readStyleOverride(CadCellStyleTemplate template)
 		{
-			var style = template.Format as TableEntity.CellStyle;
+			var style = template.Format as TableStyle.CellStyle;
 			var mapstyle = DxfClassMap.Create(style.GetType(), "TABLEFORMAT_STYLE");
-			var mapformat = DxfClassMap.Create(typeof(TableEntity.ContentFormat), "TABLEFORMAT_BEGIN");
+			var mapformat = DxfClassMap.Create(typeof(TableStyle.ContentFormat), "TABLEFORMAT_BEGIN");
 
 			this._reader.ReadNext();
 
 			bool end = false;
-			TableEntity.CellEdgeFlags currBorder = TableEntity.CellEdgeFlags.Unknown;
+			var currBorder = TableStyle.CellEdgeFlags.Unknown;
 			while (this._reader.DxfCode != DxfCode.Start)
 			{
 				switch (this._reader.Code)
 				{
 					case 95:
-						currBorder = (TableEntity.CellEdgeFlags)this._reader.ValueAsInt;
+						currBorder = (TableStyle.CellEdgeFlags)this._reader.ValueAsInt;
 						break;
 					case 1 when this._reader.ValueAsString.Equals("TABLEFORMAT_BEGIN", StringComparison.InvariantCultureIgnoreCase):
 						break;
 					case 300 when this._reader.ValueAsString.Equals("CONTENTFORMAT", StringComparison.InvariantCultureIgnoreCase):
-						readContentFormat(new CellContentFormatTemplate(new TableEntity.ContentFormat()));
+						readContentFormat(new CellContentFormatTemplate(new TableStyle.ContentFormat()));
 						break;
 					case 301 when this._reader.ValueAsString.Equals("MARGIN", StringComparison.InvariantCultureIgnoreCase):
 						this.readCellMargin(template);
 						break;
 					case 302 when this._reader.ValueAsString.Equals("GRIDFORMAT", StringComparison.InvariantCultureIgnoreCase):
-						TableEntity.CellBorder border = new TableEntity.CellBorder(currBorder);
+						TableStyle.CellBorder border = new TableStyle.CellBorder(currBorder);
 						this.readGridFormat(template, border);
 						break;
 					case 309:
@@ -1540,9 +1541,9 @@ namespace ACadSharp.IO.DXF
 			}
 		}
 
-		private void readGridFormat(CadCellStyleTemplate template, TableEntity.CellBorder border)
+		private void readGridFormat(CadCellStyleTemplate template, TableStyle.CellBorder border)
 		{
-			var map = DxfClassMap.Create(border.GetType(), nameof(TableEntity.CellBorder));
+			var map = DxfClassMap.Create(border.GetType(), nameof(TableStyle.CellBorder));
 
 			this._reader.ReadNext();
 
@@ -1563,7 +1564,7 @@ namespace ACadSharp.IO.DXF
 						border.IsInvisible = this._reader.ValueAsBool;
 						break;
 					case 340:
-						template.BorderLineTypePairs.Add(new Tuple<TableEntity.CellBorder, ulong>(border, this._reader.ValueAsHandle));
+						template.BorderLineTypePairs.Add(new Tuple<TableStyle.CellBorder, ulong>(border, this._reader.ValueAsHandle));
 						break;
 					case 309:
 						end = this._reader.ValueAsString.Equals("GRIDFORMAT_END", StringComparison.InvariantCultureIgnoreCase);
@@ -1587,7 +1588,7 @@ namespace ACadSharp.IO.DXF
 
 		private void readCellMargin(CadCellStyleTemplate template)
 		{
-			var style = template.Format as TableEntity.CellStyle;
+			var style = template.Format as TableStyle.CellStyle;
 
 			this._reader.ReadNext();
 
@@ -1852,16 +1853,16 @@ namespace ACadSharp.IO.DXF
 					cellStyle.TextHeight = this._reader.ValueAsDouble;
 					return true;
 				case 170:
-					cellStyle.CellAlignment = (TableEntity.CellAlignmentType)this._reader.ValueAsShort;
+					cellStyle.CellAlignment = (TableStyle.CellAlignmentType)this._reader.ValueAsShort;
 					return true;
 				case 283:
 					cellStyle.IsFillColorOn = this._reader.ValueAsBool;
 					return true;
 				case 90:
-					cellStyle.Type = (TableEntity.CellStyleType)this._reader.ValueAsShort;
+					cellStyle.Type = (TableStyle.CellStyleType)this._reader.ValueAsShort;
 					return true;
 				case 91:
-					cellStyle.StyleClass = (TableEntity.CellStyleClass)this._reader.ValueAsShort;
+					cellStyle.StyleClass = (TableStyle.CellStyleClass)this._reader.ValueAsShort;
 					return true;
 				case 1:
 					//Undocumented
