+++ src/ACadSharp/IO/Templates/CadTableEntityTemplate.CadTableComponentTemplate.cs
@@ -1,6 +1,6 @@
-﻿using ACadSharp.Entities;
-using ACadSharp.Objects;
+﻿using ACadSharp.Objects;
 using static ACadSharp.Entities.TableEntity;
+using static ACadSharp.IO.Templates.CadTableStyleTemplate;
 
 namespace ACadSharp.IO.Templates;
