+++ src/ACadSharp/Objects/TableStyle.BorderType.cs
@@ -1,0 +1,17 @@
+﻿namespace ACadSharp.Objects;
+
+public partial class TableStyle
+{
+	public enum BorderType : short
+	{
+		/// <summary>
+		/// Single border line.
+		/// </summary>
+		Single = 1,
+
+		/// <summary>
+		/// Double border line.
+		/// </summary>
+		Double = 2
+	}
+}
