+++ src/ACadSharp/Objects/TableStyle.CellAlignmentType.cs
@@ -1,0 +1,27 @@
+﻿namespace ACadSharp.Objects;
+
+public partial class TableStyle
+{
+	public enum CellAlignmentType
+	{
+		None = 0,
+
+		TopLeft = 1,
+
+		TopCenter = 2,
+
+		TopRight = 3,
+
+		MiddleLeft = 4,
+
+		MiddleCenter = 5,
+
+		MiddleRight = 6,
+
+		BottomLeft = 7,
+
+		BottomCenter = 8,
+
+		BottomRight = 9,
+	}
+}
