+++ src/ACadSharp/Objects/TableStyle.CellStyleType.cs
@@ -1,14 +1,19 @@
-﻿namespace ACadSharp.Entities;
+﻿namespace ACadSharp.Objects;
 
-public partial class TableEntity
+public partial class TableStyle
 {
 	public enum CellStyleType
 	{
 		Unknown = 0,
+
 		Cell = 1,
+
 		Row = 2,
+
 		Column = 3,
+
 		FormattedTableData = 4,
+
 		Table = 5
 	}
 }
