+++ src/ACadSharp/Objects/TableStyle.MarginFlags.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace ACadSharp.Objects;
+
+public partial class TableStyle
+{
+	[Flags]
+	public enum MarginFlags
+	{
+		None = 0,
+
+		Override = 1,
+	}
+}
